+++ handler/core/src/main/java/io/knotx/fragments/handler/options/FragmentsHandlerOptions.java
@@ -15,7 +15,10 @@
  */
 package io.knotx.fragments.handler.options;
 
+import static io.knotx.fragments.handler.api.actionlog.ActionLogLevel.ERROR;
+
 import io.knotx.fragments.handler.action.ActionOptions;
+import io.knotx.fragments.handler.api.actionlog.ActionLogLevel;
 import io.knotx.fragments.task.options.TaskOptions;
 import io.vertx.codegen.annotations.DataObject;
 import io.vertx.core.json.JsonObject;
@@ -33,13 +36,16 @@ public class FragmentsHandlerOptions {
 
   private Map<String, ActionOptions> actions;
 
+  private String logLevel;
+
   public FragmentsHandlerOptions(JsonObject json) {
-    init();
+    init(json);
     FragmentsHandlerOptionsConverter.fromJson(json, this);
   }
 
-  private void init() {
+  private void init(JsonObject json) {
     this.taskKey = DEFAULT_TASK_KEY;
+    logLevel = json.getString(ActionLogLevel.CONFIG_KEY_NAME, ERROR.getLevel());
   }
 
   public JsonObject toJson() {
@@ -87,6 +93,10 @@ public class FragmentsHandlerOptions {
     return this;
   }
 
+  public String getLogLevel() {
+    return logLevel;
+  }
+
   @Override
   public boolean equals(Object o) {
     if (this == o) {
@@ -110,6 +120,7 @@ public class FragmentsHandlerOptions {
     return "FragmentsHandlerOptions{" +
         "tasks=" + tasks +
         ", actions=" + actions +
+        ", actionLogMode=" + logLevel +
         '}';
   }
 }