+++ handler/engine/src/test/java/io/knotx/fragments/engine/helpers/TestFunction.java
@@ -12,17 +12,16 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *
- * The code comes from https://github.com/tomaszmichalak/vertx-rx-map-reduce.
  */
 package io.knotx.fragments.engine.helpers;
 
+import static io.knotx.fragments.handler.api.domain.FragmentResult.ERROR_TRANSITION;
 import static io.knotx.fragments.handler.api.domain.FragmentResult.SUCCESS_TRANSITION;
 
 import io.knotx.fragments.api.Fragment;
 import io.knotx.fragments.handler.api.domain.FragmentContext;
 import io.knotx.fragments.handler.api.domain.FragmentResult;
-import io.knotx.fragments.handler.api.exception.ActionFatalException;
+import io.knotx.fragments.handler.api.exception.NodeFatalException;
 import io.reactivex.Single;
 import io.vertx.core.json.JsonObject;
 import java.util.concurrent.TimeUnit;
@@ -46,6 +45,23 @@ public interface TestFunction extends Function<FragmentContext, Single<FragmentR
     };
   }
 
+  static TestFunction successWithNodeLog(JsonObject nodeObject) {
+    return fragmentContext -> {
+      Fragment fragment = fragmentContext.getFragment();
+      FragmentResult result = new FragmentResult(fragment, SUCCESS_TRANSITION, nodeObject);
+      return Single.just(result);
+    };
+  }
+
+  static TestFunction errorWithNodeLog(JsonObject nodeLog) {
+    return fragmentContext -> {
+      Fragment fragment = fragmentContext.getFragment();
+      FragmentResult result = new FragmentResult(fragment, ERROR_TRANSITION, nodeLog);
+      return Single.just(result);
+    };
+  }
+
+
   static TestFunction failure() {
     return fragmentContext -> {
       throw new RuntimeException();
@@ -54,7 +70,7 @@ public interface TestFunction extends Function<FragmentContext, Single<FragmentR
 
   static TestFunction fatal(Fragment fragment) {
     return fragmentContext -> {
-      throw new ActionFatalException(fragment);
+      throw new NodeFatalException(fragment);
     };
   }